Revert "Prod sync 1 Feb 2023"
