--- public/manifest.json
@@ -1,5 +1,0 @@
-{
-  "name": "Sommelier Finance",
-  "description": "Access to risk-managed, multi chain strategies powered by off-chain computation",
-  "iconPath": "/assets/icon/somm.svg"
-}
