+++ src/components/_buttons/ConnectButton/index.tsx
@@ -1,23 +1,115 @@
 import * as React from "react"
-import { ButtonProps } from "@chakra-ui/react"
-import { useAccount } from "wagmi"
+import { ButtonProps, useToast } from "@chakra-ui/react"
+import { Connector, useAccount, useConnect } from "wagmi"
 import ClientOnly from "components/ClientOnly"
 import { ConnectedPopover } from "./ConnectedPopover"
-import { ConnectWalletPopover } from "./ConnectWalletPopover"
+import { BaseButton, BaseButtonProps } from "../BaseButton"
+import { MoneyWalletIcon } from "components/_icons"
+import { analytics } from "utils/analytics"
+
 export interface ConnectButtonProps
   extends Omit<ButtonProps, "children"> {
+  connector: Connector
   unstyled?: boolean
 }
 
-const ConnectButton = (props: ConnectButtonProps) => {
-  const { isConnected } = useAccount()
+const ConnectButton = ({
+  connector: c,
+  unstyled,
+  ...rest
+}: ConnectButtonProps) => {
+  const toast = useToast()
+
+  const { isConnected, address, isConnecting } = useAccount()
+  const { connect } = useConnect({
+    onError: (error) => {
+      toast({
+        title: "Connection failed!",
+        description: error.message,
+        status: "error",
+        isClosable: true,
+      })
+
+      analytics.track("wallet.connect-failed", {
+        error: error.name,
+        message: error.message,
+      })
+    },
+    onSuccess: (data) => {
+      const { account } = data
+
+      if (account && account.length) {
+        analytics.track("wallet.connect-succeeded", {
+          account,
+        })
+      }
+    },
+  })
+
+  /**
+   * - If connector is ready (window.ethereum exists), it'll detect the connector
+   *   color scheme and attempt to connect on click.
+   *
+   * - If connector is not ready (window.ethereum does not exist), it'll render
+   *   as an anchor and opens MetaMask download page in a new tab
+   */
+  const conditionalProps = React.useMemo<ButtonProps>(() => {
+    return c.ready
+      ? // connector ready props
+        {
+          onClick: () => {
+            analytics.track("wallet.connect-started")
+            connect({
+              connector: c,
+            })
+          },
+        }
+      : // connector not ready props
+        {
+          as: "a",
+          href: "https://metamask.io/download",
+          target: "_blank",
+        }
+  }, [c, connect])
+
+  // Pass custom connect button styles if unstyled prop is not passed to component
+  const styles: BaseButtonProps | false = !unstyled && {
+    p: 3,
+    bg: "surface.primary",
+    borderWidth: 1,
+    borderRadius: 12,
+    borderColor: "surface.secondary",
+    minW: "max-content",
+    fontFamily: "Haffer",
+    fontSize: 12,
+    icon: MoneyWalletIcon,
+    iconProps: {
+      bgColor: "unset",
+      borderRadius: "unset",
+      _groupHover: {
+        bgColor: "unset",
+      },
+    },
+    _hover: {
+      bg: "purple.dark",
+      borderColor: "surface.tertiary",
+    },
+  }
 
   return (
     <ClientOnly>
       {isConnected ? (
         isConnected && <ConnectedPopover />
       ) : (
-        <ConnectWalletPopover {...props} />
+        <BaseButton
+          isLoading={isConnecting}
+          key={c.id}
+          {...styles}
+          {...conditionalProps}
+          {...rest}
+        >
+          Connect Wallet
+        </BaseButton>
       )}
     </ClientOnly>
   )
