+++ src/components/_cards/BridgeCard/index.tsx
@@ -1,10 +1,16 @@
-import { Heading, Stack, Text } from "@chakra-ui/react"
-import { ExternalLinkIcon } from "components/_icons"
+import {
+  Center,
+  Heading,
+  HStack,
+  Stack,
+  Text,
+} from "@chakra-ui/react"
+import { ExternalLinkIcon, InformationIcon } from "components/_icons"
 import React from "react"
 import { TransparentCard } from "../TransparentCard"
 import { Link } from "components/Link"
 import { BridgeForm } from "components/_forms/BridgeForm"
-import { useAccount } from "wagmi"
+import { useAccount, useConnect } from "wagmi"
 import ConnectButton from "components/_buttons/ConnectButton"
 import { FormProvider, useForm } from "react-hook-form"
 import { useIsMounted } from "hooks/utils/useIsMounted"
@@ -18,6 +24,7 @@ export interface BridgeFormValues {
 export const BridgeCard: React.FC = () => {
   const isMounted = useIsMounted()
   const isLarger768 = useBetterMediaQuery("(min-width: 768px)")
+  const { connectors } = useConnect()
   const { isConnected } = useAccount()
   const methods = useForm<BridgeFormValues>()
 
@@ -26,11 +33,10 @@ export const BridgeCard: React.FC = () => {
       maxW="432px"
       w="full"
       boxShadow="purpleOutline1"
-      px={{ base: 5, md: 12 }}
+      px={12}
       pt="52px"
       pb="48px"
-      borderRadius={{ base: "32px", md: "40px" }}
-      mx={4}
+      borderRadius="40px"
     >
       <Heading as="h4" fontSize={24} mb="44px">
         Bridge
@@ -50,16 +56,33 @@ export const BridgeCard: React.FC = () => {
           <ExternalLinkIcon boxSize={3} color="purple.base" />
         </Link>
       </Text>
-      {isMounted &&
+      {isLarger768 ? (
+        isMounted &&
         (isConnected ? (
           <FormProvider {...methods}>
             <BridgeForm />
           </FormProvider>
         ) : (
           <Stack>
-            <ConnectButton unstyled />
+            {connectors.map((c) => (
+              <ConnectButton connector={c} key={c.id} unstyled />
+            ))}
           </Stack>
-        ))}
+        ))
+      ) : (
+        <Center>
+          <HStack spacing="6px">
+            <InformationIcon color="red.base" boxSize="12px" />
+            <Text
+              fontSize="xs"
+              fontWeight="semibold"
+              color="red.light"
+            >
+              Bridge not yet supported on mobile
+            </Text>
+          </HStack>
+        </Center>
+      )}
     </TransparentCard>
   )
 }
