+++ src/components/_cards/PortfolioCard/index.tsx
@@ -1,6 +1,7 @@
 import {
   BoxProps,
   Heading,
+  HStack,
   Image,
   SimpleGrid,
   Spacer,
@@ -15,7 +16,7 @@ import { VFC } from "react"
 import { TransparentCard } from "../TransparentCard"
 import { toEther } from "utils/formatCurrency"
 import BondingTableCard from "../BondingTableCard"
-import { useAccount } from "wagmi"
+import { useAccount, useConnect } from "wagmi"
 import { getTokenConfig, Token } from "data/tokenConfig"
 import { TokenAssets } from "components/TokenAssets"
 import { DepositButton } from "components/_buttons/DepositButton"
@@ -57,6 +58,8 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
   )
   const stakingEnd = useStakingEnd(cellarConfig)
 
+  const { connectors } = useConnect()
+
   const { lpToken } = useUserBalances(cellarConfig)
   const { data: lpTokenData } = lpToken
 
@@ -87,14 +90,17 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
     >
       <VStack align="stretch" spacing={8}>
         <CardStatRow
-          gap={{ base: 4, md: 8, lg: 14 }}
+          spacing={{ sm: 4, md: 8, lg: 14 }}
           align="flex-start"
           justify="flex-start"
-          direction={{ base: "column", md: "row" }}
+          direction={{ sm: "column", md: "row" }}
           wrap="wrap"
         >
           <SimpleGrid
-            templateColumns="repeat(2, max-content)"
+            templateColumns={{
+              sm: "max-content",
+              lg: "repeat(2, max-content)",
+            }}
             templateRows="1fr 1fr"
             spacing={4}
             alignItems="flex-end"
@@ -143,7 +149,10 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
                 "--"
               )}
             </CardStat> */}
-            <Stack spacing={3} direction="row">
+            <Stack
+              spacing={3}
+              direction={{ sm: "row", md: "column", lg: "row" }}
+            >
               {isMounted &&
                 (isConnected ? (
                   <>
@@ -151,12 +160,21 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
                     <WithdrawButton disabled={lpTokenDisabled} />
                   </>
                 ) : (
-                  <ConnectButton unstyled />
+                  connectors.map((c) => (
+                    <ConnectButton
+                      connector={c}
+                      key={c.id}
+                      unstyled
+                    />
+                  ))
                 ))}
             </Stack>
           </SimpleGrid>
           <SimpleGrid
-            templateColumns="repeat(2, max-content)"
+            templateColumns={{
+              sm: "max-content",
+              md: "repeat(2, max-content)",
+            }}
             templateRows="repeat(2, 1fr)"
             spacing={4}
             alignItems="flex-end"
@@ -222,18 +240,11 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
               isFuture(stakingEnd.data?.endDate) && (
                 <InnerCard
                   backgroundColor="surface.tertiary"
-                  mt={8}
-                  px={7}
-                  py={7}
+                  mt="8"
+                  px="7"
+                  py="7"
                 >
-                  <Stack
-                    flexDir={{ base: "column", md: "row" }}
-                    alignItems={{
-                      base: "flex-start",
-                      md: "center",
-                    }}
-                    gap={{ base: 0, md: 4 }}
-                  >
+                  <HStack>
                     <Image
                       src="/assets/icons/somm.png"
                       alt="sommelier logo"
@@ -264,7 +275,7 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
                           : "Program Ended"}
                       </Text>
                     </LighterSkeleton>
-                  </Stack>
+                  </HStack>
                 </InnerCard>
               )}
             {isConnected && (
